+++ src/test/java/com/devcourse/ReviewRanger/reply/application/ReplyServiceTest.java
@@ -19,8 +19,8 @@ import com.devcourse.ReviewRanger.reply.domain.Reply;
 import com.devcourse.ReviewRanger.reply.dto.request.CreateReplyRequest;
 import com.devcourse.ReviewRanger.reply.dto.request.UpdateReplyRequest;
 import com.devcourse.ReviewRanger.reply.repository.ReplyRepository;
+import com.devcourse.ReviewRanger.user.UserFixture;
 import com.devcourse.ReviewRanger.user.domain.User;
-import com.devcourse.ReviewRanger.user.service.UserFixture;
 
 @ExtendWith(MockitoExtension.class)
 class ReplyServiceTest {
